Started extracting JwtHelper
