+++ priv/react_app/src/store/actions/Users/index.js
@@ -1,0 +1,3 @@
+import * as Users from './Users'
+
+export default Users
