+++ priv/react_app/src/store/actions/index.js
@@ -1,5 +1,7 @@
 import login from './Login'
+import users from './Users'
 
 export default {
-  login
+  login,
+  users
 }
